Move examples to their own file
